Move Selectors to one directory
