--- src/components/DropdownSelector/styles.css
@@ -1,24 +1,0 @@
-.row-pick-coin {
-  display: flex;
-  justify-content: space-between;
-  height: 50px;
-  margin-top: 20px;
-  width: 90%;
-  margin-left: auto;
-  margin-right: auto;
-}
-
-.text-4xl.font-bold {
-  margin-top: 5px;
-}
-
-.row-pick-coin button {
-  background-color: var(--grey-91);
-  border-color: var(--grey-91);
-  border-radius: 40px;
-}
-
-.row-pick-coin button button {
-  background-color: transparent;
-  border: 0
-}
