+++ src/components/Selector/LabelledSelector.tsx
@@ -9,7 +9,7 @@ interface Props<T> {
   style?: React.CSSProperties;
 }
 
-function Selector<T extends { id: any; displayName: string }>(props: Props<T>) {
+function LabelledSelector<T extends { id: any; displayName: string }>(props: Props<T>) {
   const { label, items, onChange, value } = props;
 
   return (
@@ -40,4 +40,4 @@ function Selector<T extends { id: any; displayName: string }>(props: Props<T>) {
   );
 }
 
-export default Selector;
+export default LabelledSelector;
