+++ src/components/Selector/index.tsx
@@ -1,0 +1,6 @@
+import AssetSelector from "./AssetSelector";
+import DropdownSelector from "./DropdownSelector";
+import LabelledSelector from "./LabelledSelector";
+import VaultSelector from "./VaultSelector";
+
+export { AssetSelector, DropdownSelector, LabelledSelector, VaultSelector };
