+++ src/pages/bridge/Issue.tsx
@@ -2,7 +2,6 @@ import { h } from "preact";
 import { Button, Checkbox, Divider, Modal } from "react-daisyui";
 import { useCallback, useEffect, useMemo, useState } from "preact/hooks";
 import LabelledInputField from "../../components/LabelledInputField";
-import LabelledSelector from "../../components/LabelledSelector";
 import { RichIssueRequest, useIssuePallet } from "../../hooks/spacewalk/issue";
 import { useVaultRegistryPallet } from "../../hooks/spacewalk/vaultRegistry";
 import { VaultRegistryVault } from "@polkadot/types/lookup";
@@ -24,46 +23,7 @@ import { CopyableAddress, PublicKey } from "../../components/PublicKey";
 import { useSecurityPallet } from "../../hooks/spacewalk/security";
 import { VoidFn } from "@polkadot/api-base/types";
 import { DateTime } from "luxon";
-import AssetSelector from "../../components/AssetSelector";
-
-interface VaultSelectorProps {
-  vaults: VaultRegistryVault[];
-  selectedVault?: VaultRegistryVault;
-  onChange: (vault: VaultRegistryVault) => void;
-}
-
-function VaultSelector(props: VaultSelectorProps): JSX.Element {
-  const { vaults, selectedVault } = props;
-
-  const items = vaults.map((vault) => {
-    return {
-      displayName: vault.id.accountId.toString(),
-      id: vault.id,
-    };
-  });
-
-  const selectedVaultItem = selectedVault
-    ? {
-        displayName: selectedVault.id.accountId.toString(),
-        id: selectedVault.id,
-      }
-    : undefined;
-
-  return (
-    <LabelledSelector
-      items={items}
-      label="Select Vault"
-      onChange={(newItem) => {
-        const newVault = vaults.find((vault) => {
-          return vault.id === newItem.id;
-        });
-        newVault && props.onChange(newVault);
-      }}
-      value={selectedVaultItem}
-      style={{ marginTop: "8px" }}
-    />
-  );
-}
+import { AssetSelector, VaultSelector } from "../../components/Selector";
 
 interface FeeBoxProps {
   bridgedAsset?: Asset;
